+++ components/mip-access/mip-access.js
@@ -1,6 +1,6 @@
-import md5 from './mip-access-md5'
+import md5 from './md5'
 import LoginModule from './mip-login'
-import evaluateAccessExpr from './mip-access-expr'
+import evaluateAccessExpr from './expr'
 
 import './mip-access.less'
